fix : Modal.tsx 하나의 파일로 관리
